requirements.txt
google-generativeai
wikipedia
python-dotenv
pydantic
langchain
langchain-google-genai
langchain-community
duckduckgo-search
pyttsx3
ddgs
uvicorn
fastapi